+++ kernel/utils/util.h
@@ -9,7 +9,7 @@ uint32_t inl(uint16_t);
 void outb(uint16_t, uint8_t);
 void outw(uint16_t, uint16_t);
 void outl(uint16_t, uint32_t);
-char* num_to_char(uint8_t num);
+char* num_to_char(uint32_t num);
 void io_wait();
 bool strcmp(const char*, const char*);
 char* strcpy(char*, const char*, uint32_t);
